Fix the deployment error
