+++ app/api/auth/[...nextauth]/route.ts
@@ -1,55 +1,6 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import CredentialsProvider from "next-auth/providers/credentials";
-import clientPromise from "@/lib/mongodb";
-import { MongoDBAdapter } from "@next-auth/mongodb-adapter";
-import { compare } from "bcryptjs";
-import User from "@/models/User"; 
-import { connectToDB } from "@/lib/connectToDB";
-
-export const authOptions: AuthOptions = {
-  adapter: MongoDBAdapter(clientPromise),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID!,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
-    }),
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials: { email: string; password: string } | undefined) {
-        if (!credentials) throw new Error("Missing credentials");
-        await connectToDB();
-        const user = await User.findOne({ email: credentials.email });
-        if (!user) throw new Error("No user found");
-        const isValid = await compare(credentials.password, user.password);
-        if (!isValid) throw new Error("Invalid password");
-        return user;
-      },
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  pages: {
-    signIn: "/auth/signin",
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) token.id = user.id;
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id as string;
-      }
-      return session;
-    },
-  },
-};
+// app/api/auth/[...nextauth]/route.ts
+import NextAuth from "next-auth";
+import { authOptions } from "./authOptions"; // or "@/lib/authOptions"
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
