+++ app/product/[id]/page.tsx
@@ -8,13 +8,14 @@ type Params = { id: string };
 // Allow dynamic routes
 export const dynamicParams = true;
 
-const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/bagella-db`, {
-    cache: "no-store",
-  });
-  const products = await res.json();
 // Static generation for each product
 export async function generateStaticParams() {
-  return products.map((product: ProductType) => ({
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/bagella-db`, {
+    cache: "no-store",
+  });
+  const products: ProductType[] = await res.json();
+
+  return products.map((product) => ({
     id: product._id,
   }));
 }
@@ -25,20 +26,30 @@ export async function generateMetadata({
 }: {
   params: Promise<Params>;
 }): Promise<Metadata> {
-    const params = await paramsPromise;
-  const product = products.find((p: { id: string; }) => p.id === params.id);
+  const params = await paramsPromise;
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/bagella-db`, {
+    cache: "no-store",
+  });
+  const products: ProductType[] = await res.json();
+  const product = products.find((p) => p._id === params.id);
+
   return {
     title: product?.name ?? "Product Not Found",
     description: product?.description ?? "",
   };
 }
 
-
-export default async function ProductDetailPage({ 
-    params: paramsPromise,
- }: { params: Promise<Params>;
-  }) {
+export default async function ProductDetailPage({
+  params: paramsPromise,
+}: {
+  params: Promise<Params>;
+}) {
   const params = await paramsPromise;
-  const product = products.find((p: ProductType) => p._id === params.id);
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/bagella-db`, {
+    cache: "no-store",
+  });
+  const products: ProductType[] = await res.json();
+  const product = products.find((p) => p._id === params.id);
+
   return <ProductDetailClient product={product} allProducts={products} />;
 }
